Remove `User` and create wrapper around `GameRepository.Save`
Remove the `User` and associated code (repo, tests) as I can't
see a use for it. If it becomes useful later, it won't be hard
to put back in.

Create a wrapper around the repository save method for games so
that we don't need to dig into the internals of `GameInteractor`
from the outside.