Added a User layer to track connections attached to players
This is a mix of changes that I unforutately forgot to commit
when I was working on them. The biggest is to support `User`s
throughout the code. User's keep track of a connection and how it
is attached to a player. This way the grimy stuff is kept out of
the core data structures, or so the thinking goes.