WIP shifting the way users are managed
I've only changed my mind on how to do this a hundred times now. The
new plan is the following:

* new users go into a pool of users, represented by `[]*User`
* each user's connection manages a read/write so reads get passed to an
  event parser
* the hub listens for events that will move a user into a given game
* `gameInteractor`s have they're own loop for managing game state

This should be what everything looks like in the future. This commit
adds the new tracking pool and game slice to the hub. The test is
currently broken so please ignore it.