Clean up hub and move Event to its own file
Removed the `gameConnections` map as it wasn't needed with the
addition of `runningGames` and `notInGame`. I also created an
interface for hubs which will likely get used in `GameInteractor`s
so that events can be handled locally. I'll create a small hub
to deal with user's who haven't joined or created a game yet and
then move them to the `GameInteractor` once either of those events
have occurred. Doing this means that the application will not have
to go outside the `GameInteractor` just to go back into it.

Event parsing will be going into the `event.go` file but I'm not
ready for that part yet.